xtrx: remove BOOST_FOREACH and obsolete API call
All instances of BOOST_FOREACH were removed in an earlier commit, but it 
was reintroduced when xtrx support was merged. I removed BOOST_FOREACH 
again here.

Also, xtrx support fails to build because it relies on the 
xtrx_open_list function, which was removed from the xtrx API in 2019. It 
was replaced with xtrx_open_string.

I don't have xtrx hardware, so I've only tested that gr-osmosdr compiles 
with these changes.